clients/python/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bidict==0.22.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.2
yarl==1.7.2
